code/requirements.txt
streamlit
google-generativeai
pdfminer.six
python-docx
dotenv